+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `ExZample.example/1` callback
+
 ## [0.4.0] - 2020-03-14
 
 ### Added